refactor : move add line station and remove line station to sections
